+++ src/index.ts
@@ -6,3 +6,5 @@ export {
     DustModelInfo,
     DustModelVariable
 } from "./model";
+export { dustState, DustState, VectorView } from "./state";
+export { dustStateTime, DustStateTime } from "./state-time";
